Split shell QUIC queueing tests
